dashboard/requirements.txt
influxdb3-python==0.3.6
streamlit==1.22
plotly
python-dotenv